GUACAMOLE-189: Refactor GuacamoleProxyConfiguration to guacamole-ext.
